[Drupal8] IEF FieldWidget refactor
