+++ src/Plugin/Field/FieldWidget/InlineEntityFormSingle.php
@@ -8,122 +8,24 @@
 namespace Drupal\inline_entity_form\Plugin\Field\FieldWidget;
 
 use Drupal\Core\Field\FieldItemListInterface;
-use Drupal\Core\Field\WidgetBase;
+use Drupal\inline_entity_form\Plugin\Field\InlineEntityWidgetBase;
 use Drupal\Core\Form\FormStateInterface;
 
 /**
  * Single value widget.
  *
+ * @ingroup field_widget
+ *
  * @FieldWidget(
  *   id = "inline_entity_form_single",
  *   label = @Translation("Inline entity form - Single value"),
+ *   multiple_values = FALSE,
  *   field_types = {
  *     "entity_reference"
  *   },
- *   settings = {
- *     "allow_existing" = FALSE,
- *     "match_operator" = "CONTAINS",
- *     "delete_references" = FALSE,
- *     "override_labels" = FALSE,
- *     "label_singular" = "",
- *     "label_plural" = ""
- *   },
- *   multiple_values = false
  * )
  */
-class InlineEntityFormSingle extends WidgetBase {
-
-  /**
-   * {@inheritdoc}
-   */
-  public static function defaultSettings() {
-    return array(
-      "allow_existing" => FALSE,
-      "match_operator" => "CONTAINS",
-      "delete_references" => FALSE,
-      "override_labels" => FALSE,
-      "label_singular" => "",
-      "label_plural" => "",
-    );
-  }
-
-  /**
-   * Returns the settings form for the current entity type.
-   *
-   * The settings form is embedded into the IEF widget settings form.
-   * Settings are later injected into the controller through $this->settings.
-   *
-   * @param $field
-   *   The definition of the reference field used by IEF.
-   * @param $instance
-   *   The definition of the reference field instance.
-   */
-  public function settingsForm(array $form, FormStateInterface $form_state) {
-    $ief_controller = inline_entity_form_get_controller($this->fieldDefinition);
-
-    $labels = $ief_controller->labels();
-    $states_prefix = 'instance[widget][settings][type_settings]';
-
-    $element['allow_existing'] = array(
-      '#type' => 'checkbox',
-      '#title' => t('Allow users to add existing @label.', array('@label' => $labels['plural'])),
-      '#default_value' => $this->settings['allow_existing'],
-    );
-    $element['match_operator'] = array(
-      '#type' => 'select',
-      '#title' => t('Autocomplete matching'),
-      '#default_value' => $this->settings['match_operator'],
-      '#options' => array(
-        'STARTS_WITH' => t('Starts with'),
-        'CONTAINS' => t('Contains'),
-      ),
-      '#description' => t('Select the method used to collect autocomplete suggestions. Note that <em>Contains</em> can cause performance issues on sites with thousands of nodes.'),
-      '#states' => array(
-        'visible' => array(
-          ':input[name="' . $states_prefix . '[allow_existing]"]' => array('checked' => TRUE),
-        ),
-      ),
-    );
-    // The single widget doesn't offer autocomplete functionality.
-    if ($form_state->get(['widget', 'type']) == 'inline_entity_form_single') {
-      $form['allow_existing']['#access'] = FALSE;
-      $form['match_operator']['#access'] = FALSE;
-    }
-
-    $element['delete_references'] = array(
-      '#type' => 'checkbox',
-      '#title' => t('Delete referenced @label when the parent entity is deleted.', array('@label' => $labels['plural'])),
-      '#default_value' => $this->settings['delete_references'],
-    );
-
-    $element['override_labels'] = array(
-      '#type' => 'checkbox',
-      '#title' => t('Override labels'),
-      '#default_value' => $this->settings['override_labels'],
-    );
-    $element['label_singular'] = array(
-      '#type' => 'textfield',
-      '#title' => t('Singular label'),
-      '#default_value' => $this->settings['label_singular'],
-      '#states' => array(
-        'visible' => array(
-          ':input[name="' . $states_prefix . '[override_labels]"]' => array('checked' => TRUE),
-        ),
-      ),
-    );
-    $element['label_plural'] = array(
-      '#type' => 'textfield',
-      '#title' => t('Plural label'),
-      '#default_value' => $this->settings['label_plural'],
-      '#states' => array(
-        'visible' => array(
-          ':input[name="' . $states_prefix . '[override_labels]"]' => array('checked' => TRUE),
-        ),
-      ),
-    );
-
-    return $element;
-  }
+class InlineEntityFormSingle extends InlineEntityWidgetBase {
 
   /**
    * {@inheritdoc}
@@ -132,20 +34,5 @@ class InlineEntityFormSingle extends WidgetBase {
     $ief_id = $this->fieldDefinition;
 
     return $element;
-  }
-
-  /**
-   * {@inheritdoc}
-   */
-  public function settingsSummary() {
-    $summary = array();
-
-    $summary[] = t('Example summary');
-    /*$placeholder = $this->getSetting('placeholder');
-    if (!empty($placeholder)) {
-      $summary[] = t('Placeholder: @placeholder', array('@placeholder' => $placeholder));
-    } */
-
-    return $summary;
   }
 }
